requirements.txt
Flask==2.2.2
requests==2.22.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
Flask_Login==0.6.2
Flask_SqlAlchemy==3.0.2
Flask_Migrate==4.0.1
Flask_Restful==0.3.9
Flask_Cors==3.0.10
Pillow==9.5.0